+++ src/main.cpp
@@ -26,7 +26,7 @@ int main(int argc, char* argv[]) {
 
     app.getSceneManager().addScene("PlayScene", std::make_shared<PlayScene>());
     app.getSceneManager().addScene("MenuScene", std::make_shared<MenuScene>());
-    app.getSceneManager().setActiveScene("MenuScene");
+    app.getSceneManager().setActiveScene("PlayScene");
 
     app.run();