+++ docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### 2024-08-10
+
+- Add cards to play scene
+
 ### 2024-08-08
 
 - Add menu scene